+++ common-libs/vc-common/src/verifier/v1/index.ts
@@ -199,7 +199,8 @@ const isValidVPProof =
     getVerifySuite: GetVerifySuiteFn,
     getProofPurposeOptions?: GetVerifierProofPurposeOptionsFn,
   ): Validator =>
-  async (value, data) => {
+  async (value, data, additionalData) => {
+    const { challenge } = additionalData
     try {
       let suite
       try {
@@ -229,7 +230,7 @@ const isValidVPProof =
         suite,
         documentLoader,
         purpose: new AuthenticationProofPurpose({
-          challenge: data.proof.challenge,
+          challenge: challenge || data.proof.challenge,
           domain: data.proof.domain,
           ...(purposeOptions || {}),
         }),